2차 완료 (초기화)
