--- src/main/java/lotto/Runner.java
@@ -1,12 +1,0 @@
-package lotto;
-
-import lotto.controller.LottoController;
-import lotto.service.LottoService;
-
-public class Runner {
-
-    public void run(LottoService lottoService){
-        LottoController lottoController=new LottoController(lottoService);
-        lottoController.doLotto();
-    }
-}
