--- src/main/java/lotto/controller/Message.java
@@ -1,11 +1,0 @@
-package lotto.controller;
-
-public class Message {
-    public static final String REQUEST_PRICE_TO_PAY="구입 금액을 입력해 주세요.";
-    public static final String PRINT_THE_NUMBER_OF_LOTTO="%d개를 구매했습니다.";
-    public static final String REQUEST_TARGET_NUMBER="당첨 번호를 입력해 주세요";
-    public static final String REQUEST_BONUS_NUMBER="보너스 번호를 입력해 주세요";
-    public static final String LOTTO_RESULT="당첨 통계\n---";
-    public static final String LOTTO_RESULT_DETAIL="%s - %d개";
-    public static final String LOTTO_RESULT_TOTAL="총 수익률은 %s%%입니다.";
-}
