+++ src/test/java/lotto/LottoTest.java
@@ -1,6 +1,5 @@
 package lotto;
 
-import lotto.domain.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
